HBOK
There seems to be no way to build this using just my repo without
snapcraft crawling back to go-ble. I have already renamed every
reference to go-ble, but it seems not to have helped. I suspect it might
be crawling up into the .git and traversing to the repo it was cloned
from from there.